Laser cutter is down
I noticed the back of the laser cutter was open and the bed is out on to the side. I am not sure what the history is and I tried chatting on Slack but got no response. I am assuming whoever opened up the back of the laser just forgot to down it here