fixed wokny file format
So, this was written in LibreOffice and the file extension was just
changed to `.md`. So, the actual file was all the guts of an `.odt`.
Unfortunately, you can't just save it as a `.txt` from within
LibreOffice. When you type a quote (") in LibreOffice (or most any
WYSIWYG editor), it inserts either an open-quote or a close-quote,
depending on the context. These characters are supported in UTF-8, so
they stay in the file when you save it to text.

Moral of the story: just write stuff in a *text editor*, not a word
processor.