a system for linking logs and a layout for SOPs
To link to a maintenance log, add "log" data to one of the items in
_data/machines.yml
See the entry for the QD-1390 for an example.
To get that to appear in an SOP, the title of the SOP (in its YAML
header) must match its name in the machines.yml file. It must also use
the SOP layout.
In the future, I would like the title of the SOP to get colored
according to the machine's status, but the "machine" div classes do
wierd sizing stuff that I don't understand, so I can't just put the
title in those. It may be good enough to just throw that styling into
the SOP layout, even though that's kind of against the Jekyll paradigm.