added a "you are here" hint to the page trees
I also changed the parent page link from ".." to the title of the parent
page, since it is no longer ambiguous which page is the current one.

It might have been nice to put the "you are here hint" on the same line,
but the way I recursively generate the tree, that would require a whole
new argument to keep track of whether we're on the current page... I
could also do some ugly regex after the preformatted tree is generated,
but that doesn't seem like a good solution either. I think the slightly
awkward appearance is worth it for the sake of future maintainability.